feat : 채팅 추가 모달 구현 및 스타일 적용
