+++ components/chat/Chat.module.scss
@@ -85,10 +85,14 @@
     justify-content: center;
     align-items: center;
     color: white;
-    background-color: $brand-pink;
+    background-color: $light-pink;
     border: none;
     outline: none;
     border-radius: 0.5rem;
+    &:hover {
+      background-color: $brand-pink;
+      transition: 0.3s;
+    }
   }
 }
