+++ pages/chat-list/ChatList.module.scss
@@ -1,4 +1,6 @@
 @import '@/styles/index.scss';
+@import '../../styles/_variables';
+
 
 .list_container {
   @include container;
@@ -40,4 +42,23 @@
 }
 
 .chat_updated {
+}
+
+.chatPlusBtn { 
+  position: fixed;
+  width: 50px;
+  height: 50px;
+  bottom: 12rem;
+  right: 3rem;
+  background-color: $light-pink;
+  color : white;
+  border: 0;
+  font-size: 3rem;
+  border-radius: 50%;
+  &:hover {
+    background-color: $brand-pink;
+    transition: 0.3s;
+  }
+
+
 }
