+++ pages/chat-list/all.tsx
@@ -3,16 +3,16 @@ import { useEffect, useState } from 'react';
 import { useRouter } from 'next/router';
 import { Chat } from '@/@types/types';
 import Image from 'next/image';
-import CreateChat from '@/components/ChatList/CreateChat';
 import { useRecoilValue } from 'recoil';
 import { userIdState } from '@/recoil/atoms/userIdState';
+import { formattingTime, todayDate } from '@/utils/formattedTimeData';
+import ChatListModal from '@/components/ChatList/ChatListModal';
 import chatListAPI from '../../apis/chatListAPI';
 import styles from './ChatList.module.scss';
-import { formattingTime, todayDate } from '@/utils/formattedTimeData';
 
 export default function AllChatList() {
   const router = useRouter();
-
+  const [isModal, setIsModal] = useState(false);
   const [allChatList, setAllChatList] = useState<Chat[]>([]);
   const getAllChat = async () => {
     const chatAllList = await chatListAPI.getAllChatList();
@@ -44,10 +44,15 @@ export default function AllChatList() {
   const today = new Date();
   const isToday = today.toISOString().split('T')[0];
 
-
+  const handleModal = () => {
+    setIsModal(!isModal);
+  };
   return (
     <ul>
-      <CreateChat />
+       <button className={styles.chatPlusBtn} type="button" onClick={() => setIsModal(true)}>
+        +
+      </button>
+      {isModal && <ChatListModal handleModal={handleModal} />}
       {allChatList.map(chat => {
         const isincluded = chat.users.some(checkIncluded);
         const dateString = todayDate(chat.updatedAt);
